+++ game/player.py
@@ -4,7 +4,12 @@ class Player:
         self.score = 0
         self.tiles = []
 
+    def add_letter(self, letter):
+        self.tiles.append(letter)
+
+    def get_letters(self):
+        return self.tiles
+
 
     def get_score(self):
         return self.score
-