+++ game/scrabble.py
@@ -18,3 +18,31 @@ class ScrabbleGame:
         else:
             index = (self.players.index(self.current_player) + 1) % len(self.players)
             self.current_player = self.players[index]
+
+    def valid_word(self, word: str, player: Player) -> bool:
+            player_letters = player.tiles
+            hand_letters = list(player_letters)
+            for letter in word:
+                if letter in hand_letters:
+                    hand_letters.remove(letter)  
+                else:
+                    return False
+            return True
+    def can_place_word(self, word: str, location: tuple, orientation: str) -> bool:
+        x, y = location
+        len_word = len(word)
+        board = self.board.grid
+
+        if orientation == "H":
+            if x < 0 or x + len_word > len(board) or y < 0 or y >= len(board[0]):
+                return False
+        else:
+            if x < 0 or x >= len(board) or y < 0 or y + len_word > len(board[0]):
+                return False
+
+        for i in range(len_word):
+            cell = board[x + i][y] if orientation == "H" else board[x][y + i]
+            if cell.letter and cell.letter != word[i]:
+                return False
+
+        return True